DDF-5495 refactor CqlRequest
