+++ catalog/ui/catalog-ui-search-api/pom.xml
@@ -38,16 +38,6 @@
             <artifactId>platform-util</artifactId>
             <version>${project.version}</version>
         </dependency>
-        <dependency>
-            <groupId>org.codice.thirdparty</groupId>
-            <artifactId>geotools-suite</artifactId>
-            <version>${org.geotools.bundle.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.sparkjava</groupId>
-            <artifactId>spark-core</artifactId>
-            <version>${spark.version}</version>
-        </dependency>
     </dependencies>
 
     <build>
@@ -71,8 +61,7 @@
                         </Export-Package>
                         <Embed-Dependency>
                             catalog-core-api-impl,
-                            platform-util,
-                            spark-core
+                            platform-util
                         </Embed-Dependency>
                     </instructions>
                 </configuration>
